requirements.txt
# DevBridge Requirements
# Python 3.13+

# Web Framework
fastapi==0.128.0
uvicorn[standard]==0.34.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-multipart==0.0.21

# Security
bcrypt==5.0.0

# Terminal/PTY Support
pywinpty==3.0.2; sys_platform == 'win32'

# Utilities
python-dotenv==1.2.1
